[FIX] according to @hparfr review
