+++ roulier/carriers/laposte/laposte_decoder.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 """Laposte XML -> Python."""
-import email.parser
 from lxml import objectify
 from roulier.codec import Decoder
 
@@ -8,29 +7,7 @@ from roulier.codec import Decoder
 class LaposteDecoder(Decoder):
     """Laposte XML -> Python."""
 
-    def get_parts(self, response):
-        head_lines = ''
-        for k, v in response.raw.getheaders().iteritems():
-            head_lines += str(k)+':'+str(v)+'\n'
-
-        full = head_lines + response.content
-
-        parser = email.parser.Parser()
-        decoded_reply = parser.parsestr(full)
-        parts = {}
-        start = decoded_reply.get_param('start').lstrip('<').rstrip('>')
-        i = 0
-        for part in decoded_reply.get_payload():
-            cid = part.get('content-Id', '').lstrip('<').rstrip('>')
-            if (not start or start == cid) and 'start' not in parts:
-                parts['start'] = part.get_payload()
-            else:
-                parts[cid or 'Attachment%d' % i] = part.get_payload()
-            i += 1
-
-        return parts
-
-    def decode(self, response):
+    def decode(self, response, parts):
         payload_xml = response['payload']
         tag, content = self.decode_payload(payload_xml)
         if tag.endswith('getProductInterResponse'):
@@ -44,10 +21,9 @@ class LaposteDecoder(Decoder):
                 cn23_cid = False
                 content['cn23'] = False
 
-            parts_dict = self.get_parts(response['response'])
-            content['label'] = parts_dict[label_cid.replace('cid:', '')]
+            content['label'] = parts[label_cid.replace('cid:', '')]
             if cn23_cid:
-                content['cn23'] = parts_dict[cn23_cid.replace('cid:', '')]
+                content['cn23'] = parts[cn23_cid.replace('cid:', '')]
 
         return content
