+++ roulier/carriers/laposte/laposte_transport.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 """Implement laposteWS."""
 import requests
+import email.parser
 from lxml import objectify, etree
 from jinja2 import Environment, PackageLoader
 from roulier.transport import Transport
@@ -135,3 +136,25 @@ class LaposteTransport(Transport):
                 "message": "Unexpected status code from server",
                 "response": response
             }
+
+    def get_parts(self, response):
+        head_lines = ''
+        for k, v in response.raw.getheaders().iteritems():
+            head_lines += str(k)+':'+str(v)+'\n'
+
+        full = head_lines + response.content
+
+        parser = email.parser.Parser()
+        decoded_reply = parser.parsestr(full)
+        parts = {}
+        start = decoded_reply.get_param('start').lstrip('<').rstrip('>')
+        i = 0
+        for part in decoded_reply.get_payload():
+            cid = part.get('content-Id', '').lstrip('<').rstrip('>')
+            if (not start or start == cid) and 'start' not in parts:
+                parts['start'] = part.get_payload()
+            else:
+                parts[cid or 'Attachment%d' % i] = part.get_payload()
+            i += 1
+
+        return parts
